+++ packages/react-native-select-pro/src/helpers/log-error.ts
@@ -1,5 +1,6 @@
 export enum ERRORS {
     NO_ARRAY_OPTIONS = 'You must pass array in the options prop',
+    SCREEN_READER_ERROR = 'isScreenReaderEnabled error',
 }
 
 // eslint-disable-next-line no-console
