+++ packages/react-native-select-pro/src/hooks/index.ts
@@ -1,1 +1,2 @@
+export { useAccessibilityScreenReader } from './use-accessibility-screen-reader';
 export { useAnimation } from './use-animation';
